+++ nodemon.json
@@ -1,5 +1,5 @@
 {
   "watch": ["app"],
   "ext": "ts",
-  "exec": "ts-node --project tsconfig.server.json ./server/index.ts"
+  "exec": "ts-node --project tsconfig.server.json ./server/server.ts"
 }
