+++ server/server.ts
@@ -1,0 +1,11 @@
+import consola from 'consola'
+import app from './app'
+
+const start = async () => {
+  app.listen(8080, '127.0.0.1')
+  consola.ready({
+    message: `Server listening on http://127.0.0.1:8080`,
+    badge: true
+  })
+}
+start()
